cc_resizefs on FreeBSD: Fix _can_skip_ufs_resize
On FreeBSD, if a UFS has trim: (-t) or MAC multilabel: (-l) flag,
resize FS fail, because the _can_skip_ufs_resize check gets tripped up
by the missing options.

This was reported at FreeBSD Bugzilla: https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=250496
and as LP# 1901958

Rather than fixing the parser as in the patches proposed there, this
pull-request rips out all of it, and simplifies the code.
We now use `growfs -N` and check if that returns an error. If it does we
can skip the resize, because we either are at the correct size, or the
filesystem in question is broken or not UFS.